+++ GLtron/zlib/gzwrite.c
@@ -71,7 +71,7 @@ local int gz_comp(state, flush)
     gz_statep state;
     int flush;
 {
-    int ret, got;
+    long ret, got;
     unsigned have;
     z_streamp strm = &(state->strm);
